// src/interfaces/spotify/spotify-artist.interface.ts
export interface ExternalLinksItem {
  name: string;
  url: string;
}